+++ libtch/torch_api.h
@@ -36,7 +36,7 @@ tensor at_shallow_clone(tensor);
 
 int at_defined(tensor);
 int at_dim(tensor);
-void at_shape(tensor, int *);
+void at_shape(tensor, int64_t *);
 int at_scalar_type(tensor);
 
 void at_backward(tensor, int, int);
